lab_6_ver2

+++ lab6.py
@@ -1,11 +1,11 @@
 from datetime import datetime
 import os
 import lab4
-from pathlib import Path
+import lab7
 
 
 class NewInformation:
-    file_to_write = "newsfeed.txt"
+    file_to_write = 'newsfeed.txt'
     length_of_description = 30
     end_record = '-' * length_of_description
 
@@ -13,10 +13,12 @@ class NewInformation:
         self.note = note_text
         self.text_in_file = ''
         self.type_of_info = ''
+        self.file_to_write = NewInformation.file_to_write
 
     def add_new_info_in_the_file(self):
         with open(self.file_to_write, "a") as text_file:
             text_file.write(self.text_in_file + '\n\n')
+        create_files_csv()
         print('The information was added\n')
 
     def info_description(self):
@@ -43,7 +45,8 @@ class AddAd(NewInformation):
         self.expire_date = str((datetime.strptime(self.date, '%d/%m/%Y') - datetime.now()).days + 1)
         self.type_of_info = 'Private Ad '
         self.description = NewInformation.info_description(self)
-        self.text_in_file = self.description + 'Actual until: ' + self.date + ', ' + self.expire_date + ' days left.' + '\n' + self.end_record + '\n'
+        self.text_in_file = self.description + 'Actual until: ' + self.date + ', ' + self.expire_date + ' days left.' \
+                            + '\n' + self.end_record + '\n'
         self.add_new_info_in_the_file()
 
 
@@ -53,64 +56,70 @@ class AddWeatherForecast(NewInformation):
         self.date = date
         self.degrees = degrees
         self.city = city
-        self.type_of_info = 'Weather Forecast'
+        self.type_of_info = 'Weather Forecast '
         self.description = NewInformation.info_description(self)
-        self.text_in_file = self.description + 'Temperature: ' + self.degrees + ' in ' + self.city + ' Date:' + self.date + '\n' + self.end_record + '\n'
+        self.text_in_file = self.description + 'Temperature: ' + self.degrees + ' in ' + self.city + \
+                            ' Date:' + self.date + '\n' + self.end_record + '\n'
         self.add_new_info_in_the_file()
 
 
 class NewRecordsFromFile:
     default_folder = str(os.path.abspath(os.curdir))
 
-    def __init__(self, count_of_records, name_of_txt_file, folder_to_load):
+    def __init__(self, count_of_records, name_of_txt_file, folder_to_load=default_folder):
         self.name_of_txt_file = name_of_txt_file
         self.count_of_records = int(count_of_records)
-        self.full_path = ''
-        self.cnt_of_all_records = 0
         self.folder_to_load = folder_to_load
+        self.full_path = os.path.join(self.folder_to_load, self.name_of_txt_file)
+        self.cnt_of_all_records = 0
+        self.line_to_insert = ''
 
-    def load_from_file(self):
-        if self.folder_to_load == '':
-            self.full_path = Path(str(self.default_folder + '\\' + self.name_of_txt_file))
-        else:
-            self.full_path = Path(str(self.folder_to_load + self.name_of_txt_file))
-        cnt_of_records_in_file = 0
-        if NewRecordsFromFile.validation_file(self):
-            with open(NewInformation.file_to_write, 'a') as file_for_insert, open(self.full_path, 'r') as file_info_inserted:
-                self.count_of_records = NewRecordsFromFile.count_of_records_less_then_len_of_file(self)
-                while (
-                        cnt_of_records_in_file < self.count_of_records) and (self.cnt_of_all_records < sum(1 for _ in open(self.full_path, 'r'))):  # adding new records in the file
-                    cur_line_of_file = file_info_inserted.readline()
-                    if (cur_line_of_file != '\n') and ('------' not in cur_line_of_file) and cur_line_of_file != '':
-                        line_to_insert = lab4.sentence_with_capit_letters((cur_line_of_file.split('. ')))
-                        cnt_of_records_in_file += 1
-                        self.cnt_of_all_records += 1
-                    elif cur_line_of_file == '\n':
-                        line_to_insert = cur_line_of_file
-                        self.cnt_of_all_records += 1
-                    elif '----------' in cur_line_of_file:
-                        line_to_insert = cur_line_of_file.capitalize()
-                        cnt_of_records_in_file += 1
-                        self.cnt_of_all_records += 1
-                    file_for_insert.write(line_to_insert)
-                print('New data was added into file')
-        else:
-            print('Wrong file or path - data was not inserted!')
-
-    def remove_file(self):
-        if NewRecordsFromFile.validation_file(self):
-            if self.cnt_of_all_records >= sum(1 for _ in open(self.full_path, 'r')):
-                os.remove(self.full_path)
-                print(f'File {self.full_path} was deleted')
-
-    def validation_file(self):
+    def is_file_exists(self):
         if os.path.exists(self.full_path):
             return True
         else:
             return False
 
+    def len_of_reading_file(self):
+        return sum(1 for _ in open(self.full_path, 'r'))
+
+    def read_from_file(self):
+        cnt_of_records_in_file = 0
+        if NewRecordsFromFile.is_file_exists(self):
+            with open(self.full_path, 'r') as file_info_inserted:
+                self.count_of_records = NewRecordsFromFile.count_of_records_less_then_len_of_file(self)
+                while (
+                        cnt_of_records_in_file < self.count_of_records) and (
+                        self.cnt_of_all_records < self.len_of_reading_file()):
+                    cur_line_of_file = file_info_inserted.readline()
+                    if (cur_line_of_file != '\n') and ('------' not in cur_line_of_file) and cur_line_of_file != '':
+                        self.line_to_insert += lab4.sentence_with_capit_letters((cur_line_of_file.split('. ')))
+                        cnt_of_records_in_file += 1
+                        self.cnt_of_all_records += 1
+                    elif cur_line_of_file == '\n':
+                        self.line_to_insert += cur_line_of_file
+                        self.cnt_of_all_records += 1
+                    elif '----------' in cur_line_of_file:
+                        self.line_to_insert += cur_line_of_file.capitalize()
+                        cnt_of_records_in_file += 1
+                        self.cnt_of_all_records += 1
+        return self.line_to_insert
+
+    def insert_into_file(self):
+        if self.is_file_exists:
+            with open(NewInformation.file_to_write, 'a') as file_for_insert:
+                file_for_insert.write(self.read_from_file())
+                print('New data was added into file')
+            create_files_csv()
+
+    def remove_file(self):
+        if self.is_file_exists:
+            if self.cnt_of_all_records >= self.len_of_reading_file():
+                os.remove(self.full_path)
+                print(f'File {self.full_path} was deleted')
+
     def count_of_records_less_then_len_of_file(self):
-        len_of_file = sum(1 for _ in open(self.full_path, 'r'))
+        len_of_file = self.len_of_reading_file()
         if self.count_of_records * 4 >= len_of_file:
             print('WARNING! You chose more records for load than in file, the whole file will be loaded')
             self.count_of_records = len_of_file
@@ -119,8 +128,14 @@ class NewRecordsFromFile:
             return self.count_of_records * 4
 
 
+def create_files_csv():
+    file_to_prov = lab7.CsvProcessing(NewInformation.file_to_write)
+    lab7.CsvProcessing.write_to_file(file_to_prov)
+    lab7.CsvProcessing.write_to_file_letters(file_to_prov)
+
+
 def validation_date(type_of_info):
-    input_date = input(f'Please write the date for {type_of_info} in format DD/MM/YYYY')
+    input_date = input(f'Please write the date for {type_of_info} in format DD/MM/YYYY:\n')
     try:
         datetime.strptime(input_date, '%d/%m/%Y')
         return input_date
@@ -130,7 +145,7 @@ def validation_date(type_of_info):
 
 
 def validation_count_of_records_type():
-    count_of_records = input('Write count of records to load in the file:')
+    count_of_records = input('Write count of records to load in the file:\n')
     try:
         int(count_of_records)
         return abs(int(count_of_records))
@@ -143,7 +158,7 @@ def type_new_info():
     type_of_feed = ''
     while type_of_feed != '0':
         type_of_feed = input('Choose number to add info:\n' + '1 - add news\n' + '2 - add private ad\n' +
-                             '3 - add weather forecast\n' + '0 - quit\n' + 'Your choice:')
+                             '3 - add weather forecast\n' + '0 - quit\n' + 'Your choice:\n')
         if type_of_feed not in ['0', '1', '2', '3']:
             print(f'Your choice is "{type_of_feed}" and it is not an expected value, please try again')
         elif type_of_feed == '1':
@@ -159,10 +174,10 @@ def type_new_info():
             AddAd(ad_text, ad_date)
         elif type_of_feed == '3':
             print('Your choice is 3 - add weather forecast:')
-            forecast_text = input('Please type the forecast:').capitalize()
-            forecast_date = validation_date('Weather forecast')
-            forecast_city = input('Please type the city:').capitalize()
-            forecast_temperature = input('Please type temperature:')
+            forecast_text = input('Please type the forecast:\n').capitalize()
+            forecast_date = validation_date('Weather forecast ')
+            forecast_city = input('Please type the city:\n').capitalize()
+            forecast_temperature = input('Please type temperature:\n')
             AddWeatherForecast(forecast_text, forecast_city, forecast_date, forecast_temperature)
     else:
         print(f'All the information is in the file {NewInformation.file_to_write}')
@@ -172,21 +187,26 @@ def main_menu():
     print(' Hi, welcome to our news feed.')
     mode_type = ''
     while mode_type != '0':
-        mode_type = input(' Would you like to type info or load from file? 1 - type, 2 load from file, 0 - exit\n')
+        mode_type = input(' Would you like to type info or load from file? 1 - type, 2 - load from file, 0 - exit\n')
         if mode_type not in ['0', '1', '2']:
             print(f'Your choice is "{mode_type}" and it is not an expected value, please try again')
         elif mode_type == '1':
             type_new_info()
         elif mode_type == '2':
             folder_to_load_data = input(
-                f'Type path to the file(if it is empty will be a current path {NewRecordsFromFile.default_folder}):')
+                f'Type path to the file(if it is empty will be a current path {NewRecordsFromFile.default_folder}):\n')
             records_to_load = validation_count_of_records_type()
             file_name = input('Type file name:\n')
             file_to_load = NewRecordsFromFile(records_to_load, file_name, folder_to_load_data)
-            NewRecordsFromFile.load_from_file(file_to_load)
-            NewRecordsFromFile.remove_file(file_to_load)
+            if NewRecordsFromFile.is_file_exists(file_to_load):
+                NewRecordsFromFile.insert_into_file(file_to_load)
+                NewRecordsFromFile.remove_file(file_to_load)
+            else:
+                print('Wrong file or path - data was not inserted!')
     else:
         print('The program finished working.')
 
 
-main_menu()
+if __name__ == "__main__":
+    main_menu()
+
